It handles with HuskyCI API errors (#12)
* [FEATURE] Insert new files to be ignored in Git

It will avoid any errors during commits of new files.

* [FEATURE] Update README Requirements with Node

* [FEATURE] Add eslint for standard code inspection

* [REFACT] Define children in PropTypes

It is required as a lint good practice to define prop validation in children

* [FEATURE] Insert airbnb lint rules

* [FEATURE] Add lint stage and its dependencies

* [REFACT] Refact to fit in new lint rules

It's the design standard used as reference for most of the companies.

* [FIX] Improve text in README

* [FEATURE] Implement SnackBar component

It is used to generate an error message when API returns an error.
It won't interrupt user experience during errors on API calling.

* [REFACT] Use babel-eslint as a parser

It is necessary to parse React code considering its particular syntaxes.

* [STYLE] Format SnackComponent to meet lint rules.

* [FEATURE] Add SnackBar component for API failures

It will render a snackbar error with 'Service is unavailabel' message.
The code was refactored to identify when API calls return an error.
In all cases, it will render the message and stop to call the API.

* [DOC] Include new env var in .env file

It will prevent preflight error in babel-eslint package verification.
It is necessary to install babel-eslint so it eslint can find this
parser. The default babel-eslint in react-scripts is not found by
eslint.